Add Mandelbrot viewer
The mandelbrot viewer was broken on the old bmason.io site. I'm not sure
what broke it but it was very broken. I fixed it up so that it's showing
sort of properly again, although the Mandelbrot part is still really
fucked up.

I also majorly refactored the whole fractal code because it was a
horrendous mess.