Fix Mandelbrot and Julia displays
The previous commit refactored all the code that handles generating and
drawing these 2 fractals. The script had been broken already, and the
code needed to be refactored so it could more easily be debugged. In the
course of the refactor the actual functionality became even more broken.
This commit gets the script back to basic functioning.

There's still a lot of work to do here. The global namespace is still
super cluttered, and the naming is lifted straight from the equations on
Wikipedia, so it's a bit hard to understand what everything's for.