Remove Chess from projects
This project isn't that impressive. More importantly, bmason.io is
currently down, and rather than debugging the current version I just
want to deploy the new Rails version (this app). Since this content
isn't currently in the app I'm removing the link for now. It can go back
in later if I want.