Refactor Mandelbrot JS code
This is a relatively large refactor undertaken to make the file easier
to understand and also pave the way for attempting to improve
performance and functionality in the future.

All of this code was not object oriented before, and I've created a
Fractal object which can be initialized to draw a certain fractal, and
can hold references to the appropriate configuration. This allows us to
keep the configuration out of 'global' namespace while still being able
to access it on demand when we need to.

Added benefit this appears to be *slightly* faster, locally at least,
and the file is much shorter and easier to read.