Speed up rendering and add color
I sped up the drawing process by writing pixel data directly rather than
drawing 1x1 rectangles. This makes the black and white version quite
fast and makes it possible to add a colored version.

I fiddled around with the colors some, but I don't really like any of
the palettes. I'm going to have to do some research into how best to
color fractals, and that'll be a good time to add the ability to
customize the display.