Add html version of resume
This is just the PDF resume translated into HTML. There is no custom
styling yet, so the HTML and PDF versions look different, and the HTML
is ugly. The link on the header now goes to the HTML version, and from
there there's a link to the PDF version. I have a feeling this will be
better for SEO and the like.