Add dynamic notes routes and first note
This commit adds a way for me to store notes on the bmason.io server as
part of this Rails app. I'm not yet sure if this idea is good or
terrible, but it works for now.

Routes are dynamically generated for every .html.md file in
public/notes. All of these routes point to 'notes#show', which renders
the markdown file as HTML.